Add files via upload
In this project I learned how Oracle Cloud and Power BI can be used to analyze sensor data using Apache Hive. I went through a flow to understand how the raw data is first uploaded to the cloud, and then loaded to Hive tables for performing queries. Finally, the results of Hive queries are imported to Microsoft Power BI.